+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Extracted the versioning support for up and downgrades from operators ([#211]).
 - Added traits to access generic operator versions ([#211]).
+- Added init_status method that uses the status default ([#211]).
 
 ### Changed
 - `BREAKING`: Changed `Conditions` trait return value to not optional ([#211]).